To railway

+++ app/Http/Controllers/JobController.php
@@ -1,8 +1,14 @@
 <?php
 
 namespace App\Http\Controllers;
+
+use App\Mail\JobPosted;
 use App\Models\Job;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
@@ -25,15 +31,18 @@ class JobController extends Controller
             'salary' => ['required']
          ]);
      
-         Job::create([
-             'title' => request('title'),
+         $job = Job::create([
+            'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
          ]);
-         return redirect('/jobs'); 
-    }
+              Mail::to($job->employer->user)->queue(
+                  new JobPosted($job)
+           );
+        return redirect('/jobs'); 
+    } 
     public function edit (Job $job) {
-        return view('jobs.edit', ['job' => $job]);
+      return view('jobs.edit', ['job' => $job]);
     }
     public function update (Job $job) {
         request()->validate([
@@ -48,7 +57,8 @@ class JobController extends Controller
         return redirect('/jobs/' . $job->id);
     }
     public function destory (Job $job) {
-        $job->delete();
+           Gate::authorize('edit-job', $job);
+           $job->delete();
         return redirect('/jobs');
     }
 }
